projects/claude_generated/python/p_18/p_18_requirements_claude.txt
yfinance==0.2.48
pandas==2.2.3
python-dotenv==1.0.1
schedule==1.2.2
